# join: improve performance