@sylvestre 2022-02-07
wahou, terrific. Is it possible to add some tests to cover the uncovered lines? (mostly error mgmt)