@jtracey 2022-02-07
Hmm... I can definitely add a test for a failure on `--check-order`, that probably should have been there before this PR anyway, but most of the new untested lines are for handling the IOError case. Does the testing framework have a good way to consistently cause reading an open file or writing to stdout to fail? Off the top of my head, the only way I know to cause that in the real world (without some sort of file system failure or kernel bug) is to use network sockets, which seems a bit heavy-handed here.