@jtracey 2022-02-08
Okay, I found some other tests that use `/dev/full` to break stdout, that should hopefully get us most of the way there. That and `--check-order` are now tested. I still don't know how to create bad reads on command.

I also added a commit to make sure when using `--check-order` that stdout gets flushed before printing the final error message.